requirements.txt
pandas
numpy
scikit-learn
shap
streamlit
joblib
matplotlib
pillow
pytesseract
opencv-python-headless
